+++ ingenialink/ethercat/servo.py
@@ -28,7 +28,7 @@ from ingenialink.constants import (
 )
 from ingenialink.dictionary import CanOpenObject, Interface
 from ingenialink.ethercat.register import EthercatRegister
-from ingenialink.exceptions import ILEcatStateError, ILError, ILIOError
+from ingenialink.exceptions import ILEcatStateError, ILError, ILIOError, ILRegisterAccessError
 from ingenialink.pdo import PDOMap, PDOServo, RPDOMap, TPDOMap
 
 logger = ingenialogger.get_logger(__name__)
@@ -204,8 +204,10 @@ class EthercatServo(EthercatServoBase):
             ILIOError,
         ) as e:
             self._handle_sdo_exception(reg, SdoOperationMsg.READ, e)
-        except AttributeError:
-            raise ILIOError(f"Error reading {reg.identifier}. The drive has been disconnected.")
+        except (AttributeError, ILError) as e:
+            raise ILIOError(
+                f"Error reading {reg.identifier}. The drive has been disconnected."
+            ) from e
         finally:
             self._lock.release()
         return value
@@ -233,8 +235,10 @@ class EthercatServo(EthercatServoBase):
             ILIOError,
         ) as e:
             self._handle_sdo_exception(reg, SdoOperationMsg.WRITE, e)
-        except AttributeError:
-            raise ILIOError(f"Error writing {reg.identifier}. The drive has been disconnected.")
+        except (AttributeError, ILError) as e:
+            raise ILIOError(
+                f"Error writing {reg.identifier}. The drive has been disconnected."
+            ) from e
         finally:
             self._lock.release()
 
@@ -249,17 +253,17 @@ class EthercatServo(EthercatServoBase):
             exception: The exception that occurred while reading or writing.
 
         Raises:
-            ILIOError: If the register cannot be read or written.
-            ILIOError: If the slave fails to acknowledge the command.
-            ILIOError: If the working counter value is wrong.
+            ILRegisterAccessError: If the register cannot be read or written.
+            ILRegisterAccessError: If the slave fails to acknowledge the command.
+            ILRegisterAccessError: If the working counter value is wrong.
         """
         default_error_msg = f"Error {operation_msg.value} {reg.identifier}"
         if isinstance(exception, pysoem.WkcError):
             wkc_errors = {
-                self.NO_RESPONSE_WORKING_COUNTER: "The working counter remained unchanged.",
-                self.NOFRAME_WORKING_COUNTER: "No frame.",
-                self.TIMEOUT_WORKING_COUNTER: "Timeout.",
-                self.UNKNOWN_FRAME_WORKING_COUNTER: "Unknown frame received.",
+                self.NO_RESPONSE_WORKING_COUNTER: "The working counter remained unchanged",
+                self.NOFRAME_WORKING_COUNTER: "No frame",
+                self.TIMEOUT_WORKING_COUNTER: "Timeout",
+                self.UNKNOWN_FRAME_WORKING_COUNTER: "Unknown frame received",
             }
             reason = wkc_errors.get(exception.wkc, f"Working counter: {exception.wkc}")
         elif isinstance(exception, (pysoem.SdoError, pysoem.MailboxError, pysoem.PacketError)):
@@ -272,7 +276,12 @@ class EthercatServo(EthercatServoBase):
             reason += error_description
         else:
             reason = str(exception)
-        raise ILIOError(f"{default_error_msg}. {reason}") from exception
+        raise ILRegisterAccessError(
+            base_message=f"{default_error_msg}. {reason}",
+            reg=reg,
+            base_exception=exception,
+            reason=reason,
+        ) from exception
 
     def _monitoring_read_data(self) -> bytes:
         """Read monitoring data frame.