+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [7.6.1] - 2026-04-29
+### Added
+- Add specialized `ILRegisterAccessError` exception for improved error handling on aborted register read/write operations
+- Create config files from DictionaryV3 defaults
+- Override configuration file values with values from another configuration file
+- Allow to upload configurations from configuration file class
+
+### Fixed
+- Fix issues of servo status listener of ethercat drives when the drive is on operational state
+
 ## [7.6.0] - 2026-03-20
 ### Added
 - Add support for tables for XDF and XCF and CSV export. XCF version has been updated to 2.2